Backend/requirements.txt
asgiref==3.6.0
channels==4.0.0
click==8.1.6
colorama==0.4.6
distlib==0.3.6
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.2
multipledispatch==0.6.0
Pillow==10.0.0
platformdirs==3.8.0
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
svgwrite==1.4.3
Tree==0.2.4
tzdata==2023.3
virtualenv==20.23.1
